+++ extensions/spotify-player/src/spotify/oauth.tsx
@@ -3,7 +3,7 @@ import fetch from "node-fetch";
 import { URLSearchParams } from "url";
 
 const clientId = "7bbb789c01ff44ed842907b7a80c404f";
-const scope = "user-library-modify user-modify-playback-state";
+const scope = "user-library-modify user-modify-playback-state user-library-read";
 //
 // user-read-currently-playing
 const oauthClient = new OAuth.PKCEClient({